+++ js/main.js
@@ -1,87 +1,3 @@
-const DESCRIPTION = [
-  'Lorem ipsum',
-  'Lorem ipsum dolor sit amet',
-  'dolor sit amet',
-  'Lorem ipsum dolor',
-  'consectetur adipiscing elit',
-  'Ut enim ad minim veniam',
-];
+import {createPhotos} from './data.js';
 
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const NAMES = [
-  'Иван',
-  'Татьяна',
-  'Петр',
-  'Елена',
-  'Мария',
-  'София',
-  'Александр',
-  'Игорь',
-  'Ксения',
-  'Тимур',
-];
-
-const SIMILAR_PHOTO_COUNT = 25;
-const SIMILAR_COMMENT_COUNT = 5;
-
-//функция возвращающая рандомное число в заданном диапазоне
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-
-  return Math.floor(result);
-};
-
-//функция возвращающая уникальное число из заданного диапазона
-const createRandomIdFromRangeGenerator = (min, max) => {
-  const previousValues = [];
-
-  return () => {
-    let currentValue = getRandomInteger(min, max);
-    if (previousValues.length >= (max - min + 1)) {
-      return null;
-    }
-    while (previousValues.includes(currentValue)) {
-      currentValue = getRandomInteger(min, max);
-    }
-    previousValues.push(currentValue);
-    return currentValue;
-  };
-};
-
-//функция возвращающая рандомный элемент из массива
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-
-const generatePhotoId = createRandomIdFromRangeGenerator(1, 25);
-const generatePhotoUrl = createRandomIdFromRangeGenerator(1, 25);
-const generateCommentId = createRandomIdFromRangeGenerator(1, 500);
-
-//функция конструктор комментария
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-  message: getRandomArrayElement(MESSAGES),
-  name: getRandomArrayElement(NAMES),
-});
-
-//функция конструктор фотографии
-const createPhoto = () => ({
-  id: generatePhotoId(),
-  url: `photos/${generatePhotoUrl()}.jpg`,
-  description: getRandomArrayElement(DESCRIPTION),
-  likes: getRandomInteger(15, 200),
-  comments: Array.from({length: SIMILAR_COMMENT_COUNT}, createComment),
-});
-
-const similarPhotos = () => Array.from({length: SIMILAR_PHOTO_COUNT}, createPhoto);
-
-similarPhotos();
+createPhotos();
